requirements.txt
# Core ML Libraries
torch
torchvision
transformers
scikit-learn
numpy
pandas

# MLOps
mlflow

# Web Framework & Server
fastapi
uvicorn

# Plotting & Visualization
matplotlib

# Data Generation
Faker
Pillow

# Utilities
tqdm
datasets
